test: extract shared RTL patterns for artwork filter specs (#13052)
* test: extract RTL-based test helpers for artwork filters

* refactor: convert SizeFilter spec to use new helpers

* refactor: convert KeywordFilter spec to use new helpers

* test: un-pend KeywordFilter specs

Follow up to https://github.com/artsy/force/pull/10598
