+++ src/Components/ArtworkFilter/ArtworkFilters/__tests__/SizeFilter.jest.tsx
@@ -1,16 +1,11 @@
-import React, { ReactElement } from "react"
-import {
-  render as originalRender,
-  RenderOptions,
-  screen,
-} from "@testing-library/react"
+import React from "react"
+import { screen } from "@testing-library/react"
 import userEvent from "@testing-library/user-event"
-
 import {
-  ArtworkFilterContextProps,
-  ArtworkFilterContextProvider,
-  useArtworkFilterContext,
-} from "Components/ArtworkFilter/ArtworkFilterContext"
+  createArtworkFilterTestRenderer,
+  currentArtworkFilterContext,
+} from "Components/ArtworkFilter/ArtworkFilters/__tests__/Utils"
+
 import {
   getPredefinedSizesByMetric,
   parseRange,
@@ -19,57 +14,13 @@ import {
   SIZES_IN_INCHES,
   getCustomSizeRangeInInches,
   CustomSize,
-} from "../SizeFilter"
+} from "Components/ArtworkFilter/ArtworkFilters/SizeFilter"
 
 jest.mock("Utils/Hooks/useMatchMedia", () => ({
   __internal__useMatchMedia: () => ({}),
 }))
 
-const render = (ui: ReactElement, options: RenderOptions = {}) =>
-  originalRender(ui, { wrapper: Wrapper, ...options })
-
-const Wrapper: React.FC = ({ children }) => {
-  return (
-    <ArtworkFilterContextProvider>
-      <ClearAllButton />
-      {children}
-      <ArtworkFilterContextInspector />
-    </ArtworkFilterContextProvider>
-  )
-}
-
-const ClearAllButton: React.FC = () => {
-  const artworkFilterContext = useArtworkFilterContext()
-
-  return (
-    <button onClick={() => artworkFilterContext.resetFilters()}>
-      Clear all
-    </button>
-  )
-}
-
-const ArtworkFilterContextInspector: React.FC = () => {
-  const artworkFilterContext = useArtworkFilterContext()
-
-  return (
-    <aside data-testid="artwork-filter-context-inspector">
-      {JSON.stringify(artworkFilterContext, null, 2)}
-    </aside>
-  )
-}
-
-const currentContext = (): ArtworkFilterContextProps => {
-  let contextInspector: HTMLElement
-  try {
-    contextInspector = screen.getByTestId("artwork-filter-context-inspector")
-  } catch (error) {
-    if (error.name === "TestingLibraryElementError")
-      throw new Error(
-        `The currentContext() helper function requires an <ArtworkFilterContextInspector /> to be mounted in the current DOM.`
-      )
-  }
-  return JSON.parse(contextInspector!.textContent!)
-}
+const render = createArtworkFilterTestRenderer()
 
 describe("SizeFilter", () => {
   it("toggles custom input render", () => {
@@ -88,13 +39,16 @@ describe("SizeFilter", () => {
 
   it("updates context on filter change", () => {
     render(<SizeFilter expanded />)
-    expect(currentContext().filters?.sizes).toEqual([])
+    expect(currentArtworkFilterContext().filters?.sizes).toEqual([])
 
     userEvent.click(screen.getAllByRole("checkbox")[0])
-    expect(currentContext().filters?.sizes).toEqual(["SMALL"])
+    expect(currentArtworkFilterContext().filters?.sizes).toEqual(["SMALL"])
 
     userEvent.click(screen.getAllByRole("checkbox")[2])
-    expect(currentContext().filters?.sizes).toEqual(["SMALL", "LARGE"])
+    expect(currentArtworkFilterContext().filters?.sizes).toEqual([
+      "SMALL",
+      "LARGE",
+    ])
   })
 
   it("updates the filter values", async () => {
@@ -107,11 +61,11 @@ describe("SizeFilter", () => {
     userEvent.type(screen.getAllByRole("spinbutton")[3], "16")
     userEvent.click(screen.getByText("Set size"))
 
-    expect(currentContext().filters?.sizes).toEqual([])
-    expect(currentContext().filters?.height).toEqual(
+    expect(currentArtworkFilterContext().filters?.sizes).toEqual([])
+    expect(currentArtworkFilterContext().filters?.height).toEqual(
       "4.724409448818897-6.299212598425196"
     )
-    expect(currentContext().filters?.width).toEqual(
+    expect(currentArtworkFilterContext().filters?.width).toEqual(
       "4.724409448818897-6.299212598425196"
     )
   })
@@ -124,19 +78,19 @@ describe("SizeFilter", () => {
     userEvent.type(screen.getAllByRole("spinbutton")[3], "24")
     userEvent.click(screen.getByText("Set size"))
 
-    expect(currentContext().filters?.sizes).toEqual([])
-    expect(currentContext().filters?.height).toEqual(
+    expect(currentArtworkFilterContext().filters?.sizes).toEqual([])
+    expect(currentArtworkFilterContext().filters?.height).toEqual(
       "4.724409448818897-9.448818897637794"
     )
-    expect(currentContext().filters?.width).toEqual("*-*")
+    expect(currentArtworkFilterContext().filters?.width).toEqual("*-*")
   })
 
   it("clears local input state after Clear All", () => {
     render(<SizeFilter expanded />)
 
     // before: filter state and input state are empty
-    expect(currentContext().filters?.height).toEqual("*-*")
-    expect(currentContext().filters?.width).toEqual("*-*")
+    expect(currentArtworkFilterContext().filters?.height).toEqual("*-*")
+    expect(currentArtworkFilterContext().filters?.width).toEqual("*-*")
     expect(screen.queryByDisplayValue("1")).not.toBeInTheDocument()
     expect(screen.queryByDisplayValue("2")).not.toBeInTheDocument()
     expect(screen.queryByDisplayValue("3")).not.toBeInTheDocument()
@@ -152,11 +106,11 @@ describe("SizeFilter", () => {
     userEvent.click(screen.getByText("Set size"))
 
     // assert: state and ui are updated
-    expect(currentContext().filters?.sizes).toEqual([])
-    expect(currentContext().filters?.width).toEqual(
+    expect(currentArtworkFilterContext().filters?.sizes).toEqual([])
+    expect(currentArtworkFilterContext().filters?.width).toEqual(
       "0.39370078740157477-0.7874015748031495"
     )
-    expect(currentContext().filters?.height).toEqual(
+    expect(currentArtworkFilterContext().filters?.height).toEqual(
       "1.1811023622047243-1.574803149606299"
     )
     expect(screen.queryByDisplayValue("1")).toBeInTheDocument()
@@ -168,8 +122,8 @@ describe("SizeFilter", () => {
     userEvent.click(screen.getByText("Clear all"))
 
     // assert: state and ui are cleared
-    expect(currentContext().filters?.height).toEqual("*-*")
-    expect(currentContext().filters?.width).toEqual("*-*")
+    expect(currentArtworkFilterContext().filters?.height).toEqual("*-*")
+    expect(currentArtworkFilterContext().filters?.width).toEqual("*-*")
     expect(screen.queryByDisplayValue("1")).not.toBeInTheDocument()
     expect(screen.queryByDisplayValue("2")).not.toBeInTheDocument()
     expect(screen.queryByDisplayValue("3")).not.toBeInTheDocument()
